Speeding up the second iteration part of the mapping software by transforming the list to np.arrays
online descriptions inform that an np.array can be handled 30 times faster than a list. Possibly it would be worth to figure out how to append the images into the np.array immediately instead of adding it to a list and only at the end transforming it to an array.